+++ packages/xcm-cfg/src/chains/chains.spec.ts
@@ -42,7 +42,7 @@ describe('chains config', () => {
           .filter((c) => h160AddressSpaceOnly(c))
           .map((c) => c.key)
           .sort()
-      ).toEqual(['darwinia', 'moonbeam', 'mythos'].sort());
+      ).toEqual(['darwinia', 'laos', 'moonbeam', 'mythos'].sort());
     });
     it('should match given chains with substrate only address space support', async () => {
       expect(
@@ -67,6 +67,7 @@ describe('chains config', () => {
           'subsocial',
           'unique',
           'zeitgeist',
+          'energywebx',
         ].sort()
       );
     });