+++ packages/xcm-cfg/src/configs/polkadot/hydration/index.ts
@@ -21,6 +21,7 @@ import {
   intr,
   kilt,
   ksm,
+  laos,
   ldot,
   myth,
   nodl,
@@ -46,6 +47,7 @@ import {
   weth_awh,
   weth_mwh,
   wud,
+  ewt,
 } from '../../../assets';
 import {
   acala,
@@ -59,6 +61,7 @@ import {
   hydration,
   interlay,
   kilt_chain,
+  laos_chain,
   moonbeam,
   mythos,
   nodle,
@@ -69,6 +72,7 @@ import {
   unique,
   zeitgeist,
   polkadotCex,
+  energywebx,
 } from '../../../chains';
 import { ExtrinsicBuilder, XcmTransferType } from '../../../builders';
 
@@ -267,6 +271,10 @@ const toCrust: AssetRoute[] = [toTransferTemplate(cru, crust, 0.04)];
 
 const toKilt: AssetRoute[] = [toTransferTemplate(kilt, kilt_chain, 0.02)];
 
+const toLaos: AssetRoute[] = [toTransferTemplate(laos, laos_chain, 0.1)];
+
+const toEnergywebx: AssetRoute[] = [toTransferTemplate(ewt, energywebx, 0.02)];
+
 const toPendulum: AssetRoute[] = [toTransferTemplate(pen, pendulum, 1.1)];
 
 const toDarwinia: AssetRoute[] = [toTransferTemplate(ring, darwinia, 4)];
@@ -349,6 +357,8 @@ export const hydrationConfig = new ChainRoutes({
     ...toEthereumViaWormhole,
     ...toInterlay,
     ...toKilt,
+    ...toLaos,
+    ...toEnergywebx,
     ...toMoonbeam,
     ...toMythos,
     ...toNodle,