+++ packages/xcm-cfg/src/configs/polkadot/index.ts
@@ -20,6 +20,8 @@ import { kiltConfig } from './kilt';
 import { pendulumConfig } from './pendulum';
 import { darwiniaConfig } from './darwinia';
 import { ajunaConfig } from './ajuna';
+import { laosConfig } from './laos';
+import { energywebxConfig } from './energywebx';
 
 export const polkadotChainsConfig: ChainRoutes[] = [
   acalaConfig,
@@ -44,4 +46,6 @@ export const polkadotChainsConfig: ChainRoutes[] = [
   kiltConfig,
   pendulumConfig,
   darwiniaConfig,
+  laosConfig,
+  energywebxConfig,
 ];