+++ packages/xcm-core/src/evm/EvmClient.ts
@@ -39,7 +39,12 @@ export class EvmClient {
     const withFallback = this.rpcs.map((rpc) => http(rpc));
     return createPublicClient({
       chain: this.chain,
-      transport: this.rpcs.length > 0 ? fallback(withFallback) : http(),
+      transport:
+        this.rpcs.length > 0
+          ? fallback(withFallback, {
+              rank: false,
+            })
+          : http(),
     });
   }