+++ packages/xcm-core/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @galacticcouncil/xcm-core
 
+## 6.6.0
+
+### Minor Changes
+
+- Support getQuote fallback price
+- Disable rank ping in EvmClient
+
+## 6.5.0
+
+### Minor Changes
+
+- Support opt metadata in SubstrateApis
+- 0d5426a: Support api metadata
+
+## 6.4.0
+
+### Minor Changes
+
+- Update ABI's
+- EvmClient fallback ranking alg
+
 ## 6.3.0
 
 ### Minor Changes